refactor(daemon): isolate restart health types
